+++ SOS/Plush/Plush.ExecInstruction.cpp
@@ -11,7 +11,7 @@ namespace Plush
 	// Add an instruction to the instruction set for this individual.
 	inline size_t exec_enable_instruction(Environment& _env)
 	{
-		int n = _env.pop<long>();	// Instruction index
+		long n = _env.pop<long>();	// Instruction index
 
 		_env.enable_function(n);
 
@@ -21,13 +21,13 @@ namespace Plush
 	// Add instructions to the instruction set for this individual using the integers in the integer stack.
 	inline size_t exec_enable_instructions(Environment& _env)
 	{
-		int number_of_instructions_to_add = _env.pop<long>();	// Number of instructions to add
-		int i = _env.pop<long>();								// Instruction index
+		long number_of_instructions_to_add = _env.pop<long>();	// Number of instructions to add
+		long i = _env.pop<long>();								// Instruction index
 
 		if ((number_of_instructions_to_add > 0) && (_env.length<long>() >= (number_of_instructions_to_add - 1)))
 		{
 			_env.enable_function(i);
-			int n = number_of_instructions_to_add - 1;
+			long n = number_of_instructions_to_add - 1;
 
 			while (n > 0)
 			{
@@ -48,7 +48,7 @@ namespace Plush
 	// Remove instruction from the instruction set for this individual.
 	inline size_t exec_disable_instruction(Environment& _env)
 	{
-		int n = _env.pop<long>();	// Instruction index
+		long n = _env.pop<long>();	// Instruction index
 
 		_env.disable_function(n);
 
@@ -83,8 +83,8 @@ namespace Plush
 	// for possible access during the execution of the body of the loop
 	inline size_t exec_do_range(Environment & _env)
 	{
-		int n = _env.pop<long>();	// destination index
-		int i = _env.pop<long>();	// current index
+		long n = _env.pop<long>();	// destination index
+		long i = _env.pop<long>();	// current index
 
 		if (n == i)
 			_env.push<long>(i);
@@ -129,7 +129,7 @@ namespace Plush
 
 	inline size_t exec_do_count(Environment & _env)
 	{
-		int n = _env.pop<long>();	// destination index
+		long n = _env.pop<long>();	// destination index
 
 		if (n > 0)
 		{
@@ -144,7 +144,7 @@ namespace Plush
 
 	inline size_t exec_do_times(Environment & _env)
 	{
-		int n = _env.pop<long>();	// destination index
+		long n = _env.pop<long>();	// destination index
 
 		if (n > 0)
 		{
@@ -432,8 +432,8 @@ namespace Plush
 
 	inline size_t code_do_range(Environment & _env)
 	{
-		int n = _env.pop<long>();	// destination index
-		int i = _env.pop<long>();	// current index
+		long n = _env.pop<long>();	// destination index
+		long i = _env.pop<long>();	// current index
 
 		if (n == i)
 		{
@@ -476,7 +476,7 @@ namespace Plush
 
 	inline size_t code_do_count(Environment & _env)
 	{
-		int n = _env.pop<long>();	// destination index
+		long n = _env.pop<long>();	// destination index
 
 		if (n > 0)
 		{
@@ -491,7 +491,7 @@ namespace Plush
 
 	inline size_t code_do_times(Environment & _env)
 	{
-		int n = _env.pop<long>();	// destination index
+		long n = _env.pop<long>();	// destination index
 
 		if (n > 0)
 		{
@@ -507,7 +507,7 @@ namespace Plush
 
 	inline size_t code_extract(Environment & _env)
 	{
-		unsigned int item_number = std::abs(_env.pop<long>());	// index
+		long item_number = std::abs(_env.pop<long>());	// index
 
 		if (item_number != 0)
 		{
@@ -523,7 +523,7 @@ namespace Plush
 			if (number_of_items > 0)
 			{
 				// Take modulo the number of blocks to ensure that it is within the meaningful range.
-				int n = item_number - 1;
+				long n = item_number - 1;
 				item_number = std::abs(n) % number_of_items + 1;
 
 				Genome_section<CodeAtom> sub_block = genome.get_subitem(item_number);
@@ -600,7 +600,7 @@ namespace Plush
 
 	inline size_t code_insert(Environment & _env)
 	{
-		unsigned int item_number = std::abs(_env.pop<long>());	// index
+		long item_number = std::abs(_env.pop<long>());	// index
 
 		// Get reference to genome stack
 		Genome<CodeAtom>& genome = _env.get_stack<CodeAtom>();
@@ -617,7 +617,7 @@ namespace Plush
 		// Take modulo the number of blocks to ensure that it is within the meaningful range.
 		if (item_number != 0)
 		{
-			int n = item_number - 1;
+			long n = item_number - 1;
 			item_number = std::abs(n) % number_of_items + 1;
 			item_number = (item_number == number_of_items) ? 0 : item_number;
 		}
@@ -703,7 +703,7 @@ namespace Plush
 
 	inline size_t code_nth(Environment & _env)
 	{
-		unsigned int index = std::abs(_env.pop<long>());	// index
+		long index = std::abs(_env.pop<long>());	// index
 
 		// Get reference to genome stack
 		Genome<CodeAtom>& genome = _env.get_stack<CodeAtom>();
@@ -732,7 +732,7 @@ namespace Plush
 
 	inline size_t code_nthcdr(Environment & _env)
 	{
-		unsigned int item_number = std::abs(_env.pop<long>());	// index
+		long item_number = std::abs(_env.pop<long>());	// index
 
 		// Get reference to genome stack
 		Genome<CodeAtom>& genome = _env.get_stack<CodeAtom>();
